Backup incomplete code
This is an attempt to implement Schema with arena or Rc so that to make
its fields can reference to each other to form a graph instead of a
tree. But after trying arena / Rc, I find that a GC would be better.
Otherwise, the recursive definiton / similar schema merging can be
implemented after constructing the schema, thus requiring no
consideration on the reference problem.